update xml attributes for getMusicFolders
