+++ core/types/subsonic.go
@@ -106,49 +106,3 @@ type SubsonicTokenInfo struct {
 type SubsonicTokenInfoResponse struct {
 	SubsonicResponse SubsonicTokenInfo `json:"subsonic-response"`
 }
-
-type OpenSubsonicExtensions struct {
-	Name     string `xml:"name" json:"name"`
-	Versions []int  `xml:"versions" json:"versions"`
-}
-
-type SubsonicOpenSubsonicExtensions struct {
-	XMLName                xml.Name                  `xml:"subsonic-response" json:"-"`
-	Xmlns                  string                    `xml:"xmlns,attr" json:"-"`
-	Status                 string                    `xml:"status,attr" json:"status"`
-	Version                string                    `xml:"version,attr" json:"version"`
-	Type                   string                    `xml:"type,attr" json:"type"`
-	ServerVersion          string                    `xml:"serverVersion,attr" json:"serverVersion"`
-	OpenSubsonic           bool                      `xml:"openSubsonic,attr" json:"openSubsonic"`
-	Error                  *SubsonicError            `xml:"error,omitempty" json:"error,omitempty"`
-	OpenSubsonicExtensions []*OpenSubsonicExtensions `xml:"openSubsonicExtensions" json:"openSubsonicExtensions"`
-}
-
-type SubsonicOpenSubsonicExtensionsResponse struct {
-	SubsonicResponse SubsonicOpenSubsonicExtensions `json:"subsonic-response"`
-}
-
-type MusicFolders struct {
-	MusicFolder []MusicFolder `xml:"musicFolder" json:"musicFolder"`
-}
-
-type MusicFolder struct {
-	Id   int    `xml:"id" json:"id"`
-	Name string `xml:"name" json:"name"`
-}
-
-type SubsonicMusicFolders struct {
-	XMLName       xml.Name       `xml:"subsonic-response" json:"-"`
-	Xmlns         string         `xml:"xmlns,attr" json:"-"`
-	Status        string         `xml:"status,attr" json:"status"`
-	Version       string         `xml:"version,attr" json:"version"`
-	Type          string         `xml:"type,attr" json:"type"`
-	ServerVersion string         `xml:"serverVersion,attr" json:"serverVersion"`
-	OpenSubsonic  bool           `xml:"openSubsonic,attr" json:"openSubsonic"`
-	Error         *SubsonicError `xml:"error,omitempty" json:"error,omitempty"`
-	MusicFolders  *MusicFolders  `xml:"musicFolders" json:"musicFolders"`
-}
-
-type SubsonicMusicFoldersResponse struct {
-	SubsonicResponse SubsonicMusicFolders `json:"subsonic-response"`
-}
